+++ tests/test_jobstores.py
@@ -1,150 +1,153 @@
-import datetime
-import logging
-from unittest import mock
+import warnings
+from datetime import datetime
 
+import pytest
 from apscheduler import events
 from apscheduler.events import JobExecutionEvent, JobSubmissionEvent
-from django.db.backends.utils import CursorWrapper
-from django.db.utils import OperationalError
 from django.utils import timezone
 
 from django_apscheduler.jobstores import (
     DjangoJobStore,
     register_job,
+    register_events,
 )
 from django_apscheduler.models import DjangoJob, DjangoJobExecution
-from django_apscheduler.util import get_django_internal_datetime
-
-logging.basicConfig()
+from tests.conftest import DummyScheduler, dummy_job
 
 
-def test_add_job(db, job, scheduler):
-    scheduler.add_job(job, trigger="interval", seconds=1, id="job")
-
-    scheduler.start()
-
-    assert DjangoJob.objects.count() == 1
-
-    # Add job second time
-    scheduler.add_job(
-        job, trigger="interval", seconds=1, id="job", replace_existing=True
-    )
-
-    assert DjangoJob.objects.count() == 1
-
-
-def test_issue_20(db, job, scheduler):
-    scheduler.add_job(job, trigger="interval", seconds=1, id="job")
-    scheduler.start()
-
-    assert DjangoJob.objects.count() == 1
-
-    scheduler.remove_job("job")
-
-    assert DjangoJob.objects.count() == 0
-
-
-def test_remove_job(db, job, scheduler):
-    """ This test checks issue https://github.com/jarekwg/django-apscheduler/issues/6 """
-
-    scheduler.add_job(job, trigger="interval", seconds=1, id="job")
-    scheduler.start()
-
-    assert DjangoJob.objects.count() == 1
-    assert len(scheduler.get_jobs()) == 1
-
-    dbJob = DjangoJob.objects.first()
-    dbJob.delete()
-
-    assert len(scheduler.get_jobs()) == 0
-
-
-def job_for_tests():
-    job_for_tests.mock()
-
-
-job_for_tests.mock = mock.Mock()
-
-
-def test_try_add_job_then_start(db, scheduler):
-    scheduler.add_job(
-        job_for_tests, next_run_time=timezone.now(), misfire_grace_time=None,
-    )
-    scheduler.start()
-    scheduler._process_jobs()
-
-    assert job_for_tests.mock.call_count == 1
-
-
-def test_register_job_dec(db, job, scheduler):
-    register_job(scheduler, "interval", seconds=1)(job)
-
-    scheduler.start()
-
-    assert DjangoJob.objects.count() == 1
-
-    dbj = DjangoJob.objects.first()
-
-    assert dbj.id == "tests.conftest.dummy_job"
-
-    j = scheduler.get_jobs()[0]
-
-    assert j.id == "tests.conftest.dummy_job"
-
-
-def test_job_events(db, job, djangojobstore, scheduler):
-    scheduler.add_job(job, trigger="interval", seconds=1, id="job")
-    scheduler.start()
-
-    dj = DjangoJob.objects.last()
-    dj.next_run_time -= datetime.timedelta(seconds=2)
-    dj.save()
-
-    now = timezone.now()
-    scheduler._dispatch_event(
-        JobSubmissionEvent(events.EVENT_JOB_SUBMITTED, "job", djangojobstore, [now])
-    )
-    scheduler._dispatch_event(
-        JobExecutionEvent(events.EVENT_JOB_EXECUTED, "job", djangojobstore, now)
-    )
-
-    assert DjangoJobExecution.objects.count() == 1
-
-
-def test_issue_15(db, djangojobstore):
-    """
-    This test covers bug from https://github.com/jarekwg/django-apscheduler/issues/15
-    """
-    srt = timezone.now()
-
-    job = DjangoJob.objects.create(id="test", next_run_time=timezone.now())
-    DjangoJobExecution.objects.create(
-        job=job, run_time=get_django_internal_datetime(srt)
-    )
-
-    djangojobstore.handle_submission_event(
-        mock.Mock(
-            spec=JobSubmissionEvent,
-            job_id=job.id,
-            code=events.EVENT_JOB_SUBMITTED,
-            scheduled_run_times=[srt],
-        )
-    )
-
-
-def test_reconnect_on_db_error(transactional_db):
-    counter = [0]
-
-    def mocked_execute(self, *a, **k):
-        counter[0] += 1
-
-        if counter[0] == 1:
-            raise OperationalError()
-        else:
-            return []
-
-    with mock.patch.object(CursorWrapper, "execute", mocked_execute):
+class TestDjangoResultStoreMixin:
+    def test_start_gets_scheduler_lock(self):
         store = DjangoJobStore()
-        # DjangoJob.objects._last_ping = 0
 
-        assert store.get_due_jobs(now=timezone.now()) == []
+        store.start(DummyScheduler(), "djangojobstore")
+        assert store.lock is not None
+
+    @pytest.mark.django_db
+    def test_handle_submission_event_not_supported_raises_exception(self, jobstore):
+        event = JobSubmissionEvent(
+            events.EVENT_ALL, "test_job", jobstore, [timezone.now()]
+        )
+
+        with pytest.raises(NotImplementedError):
+            jobstore.handle_submission_event(event)
+
+    @pytest.mark.django_db
+    def test_handle_submission_event_creates_job_execution(
+        self, jobstore, create_add_job
+    ):
+        job = create_add_job(jobstore, dummy_job, datetime(2016, 5, 3))
+        event = JobSubmissionEvent(
+            events.EVENT_JOB_SUBMITTED, job.id, jobstore, [timezone.now()]
+        )
+        jobstore.handle_submission_event(event)
+
+        assert DjangoJobExecution.objects.filter(job_id=event.job_id).exists()
+
+    @pytest.mark.django_db
+    def test_handle_execution_event_not_supported_raises_exception(self, jobstore):
+        event = JobExecutionEvent(
+            events.EVENT_ALL, "test_job", jobstore, timezone.now()
+        )
+
+        with pytest.raises(NotImplementedError):
+            jobstore.handle_execution_event(event)
+
+    @pytest.mark.django_db
+    def test_handle_execution_event_creates_job_execution(
+        self, jobstore, create_add_job
+    ):
+        job = create_add_job(jobstore, dummy_job, datetime(2016, 5, 3))
+        event = JobExecutionEvent(
+            events.EVENT_JOB_EXECUTED, job.id, jobstore, timezone.now()
+        )
+        jobstore.handle_execution_event(event)
+
+        assert DjangoJobExecution.objects.filter(job_id=event.job_id).exists()
+
+    @pytest.mark.django_db
+    def test_handle_error_event_not_supported_raises_exception(self, jobstore):
+        event = JobExecutionEvent(
+            events.EVENT_ALL, "test_job", jobstore, timezone.now()
+        )
+
+        with pytest.raises(NotImplementedError):
+            jobstore.handle_error_event(event)
+
+    @pytest.mark.django_db
+    @pytest.mark.parametrize(
+        "event_code",
+        [
+            events.EVENT_JOB_MAX_INSTANCES,
+            events.EVENT_JOB_MISSED,
+            events.EVENT_JOB_ERROR,
+        ],
+    )
+    def test_handle_error_event_creates_job_execution(
+        self, jobstore, create_add_job, event_code
+    ):
+        job = create_add_job(jobstore, dummy_job, datetime(2016, 5, 3))
+        event = JobExecutionEvent(event_code, job.id, jobstore, timezone.now())
+        jobstore.handle_error_event(event)
+
+        assert DjangoJobExecution.objects.filter(job_id=event.job_id).exists()
+
+    @pytest.mark.django_db
+    def test_handle_error_event_no_exception_sets_exception_text(
+        self, jobstore, create_add_job
+    ):
+        job = create_add_job(jobstore, dummy_job, datetime(2016, 5, 3))
+        event = JobExecutionEvent(
+            events.EVENT_JOB_ERROR, job.id, jobstore, timezone.now()
+        )
+        jobstore.handle_error_event(event)
+
+        ex = DjangoJobExecution.objects.get(job_id=event.job_id)
+
+        assert "raised an error!" in ex.exception
+
+    @pytest.mark.django_db
+    def test_register_event_listeners_registers_listeners(self, jobstore):
+        jobstore.register_event_listeners()
+        registered_event_codes = [event[1] for event in jobstore._scheduler._listeners]
+
+        assert all(
+            event_code in registered_event_codes
+            for event_code in [
+                events.EVENT_JOB_SUBMITTED,
+                events.EVENT_JOB_EXECUTED,
+                events.EVENT_JOB_MAX_INSTANCES
+                | events.EVENT_JOB_ERROR
+                | events.EVENT_JOB_MISSED,
+            ]
+        )
+
+
+class TestDjangoJobStore:
+    """
+    We use the APScheduler tests to verify that DjangoJobStore implements the interface correctly.
+
+    This test class should only contain tests that are specific to DjangoJobStore
+
+    See 'test_apscheduler_jobstore.py' for details
+    """
+
+    pass
+
+
+@pytest.mark.django_db
+def test_register_events_raises_deprecation_warning(scheduler, jobstore):
+
+    with warnings.catch_warnings(record=True) as w:
+
+        register_events(scheduler, jobstore)
+        assert len(w) == 1
+        assert issubclass(w[-1].category, DeprecationWarning)
+        assert "deprecated" in str(w[-1].message)
+
+
+@pytest.mark.django_db
+def test_register_job(scheduler, jobstore):
+    register_job(scheduler, "interval", seconds=1)(dummy_job)
+    scheduler.start()
+
+    assert DjangoJob.objects.count() == 1